% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{The frequency of words}
\usage{
freq(x)
}
\arguments{
\item{x}{A character vector of words.}
}
\value{
A data frame with \code{char} and \code{freq} columns.
}
\description{
This function returns the frequency of words.
}
