% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\alias{stopwords_cn}
\alias{stopwords_en}
\alias{stopwords_full}
\title{Built-in stopword lists}
\format{
Three character vectors:
\describe{
\item{stopwords_cn}{A compiled Chinese stopword list. The upstream list
also contains some punctuation, numbers, and English entries.}
\item{stopwords_en}{A commonly used English stopword list.}
\item{stopwords_full}{A combined Chinese and English stopword list.}
}
}
\source{
The source files were downloaded from commit
\verb{0f77b15} of
\url{https://github.com/Northriven/Stopwords} on 2026-08-05:
\itemize{
\item \url{https://github.com/Northriven/Stopwords/blob/0f77b15d8658cdb14cd12f0338af014c902c9d40/stopwords_cn.txt}
\item \url{https://github.com/Northriven/Stopwords/blob/0f77b15d8658cdb14cd12f0338af014c902c9d40/stopwords_english.txt}
\item \url{https://github.com/Northriven/Stopwords/blob/0f77b15d8658cdb14cd12f0338af014c902c9d40/stopwords_full.txt}
}
}
\usage{
data(stopwords_cn)

data(stopwords_en)

data(stopwords_full)
}
\description{
Three UTF-8 character vectors of stopwords bundled with \code{jiebaRS}. They are
provided for explicit use with the \code{stop_word} argument; none is enabled by
default.
}
\examples{
cutter <- worker(stop_word = stopwords_cn)
segment("\u8fd9\u662f\u4e00\u4e2a\u6d4b\u8bd5", cutter)

}
