% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobin.R
\name{tobin}
\alias{tobin}
\title{Convert a simhash value to binary}
\usage{
tobin(x)
}
\arguments{
\item{x}{A non-missing character scalar containing a decimal unsigned
64-bit integer.}
}
\value{
A character scalar containing the 64-bit binary representation.
}
\description{
Convert a decimal simhash value into a fixed-width 64-bit binary string.
}
\examples{
tobin("1")
tobin("2")
}
