% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{keywords_df}
\alias{keywords_df}
\title{Extract keywords as a data frame}
\usage{
keywords_df(x, jiebar)
}
\arguments{
\item{x}{A character to analyze.}

\item{jiebar}{A \code{jieba_worker} object created with \code{worker(type = "keywords")}.}
}
\value{
A data frame with \code{term} and \code{weight} columns.
}
\description{
Convenience wrapper around \code{\link[=keywords]{keywords()}} that always returns a data frame with
\code{term} and \code{weight} columns.
}
