% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-dict.R
\name{new_user_word}
\alias{new_user_word}
\title{Add user word}
\usage{
new_user_word(worker, words, tags = "n")
}
\arguments{
\item{worker}{A \code{jieba_worker} object.}

\item{words}{A single string or a character vector of new words.}

\item{tags}{A single tag or a character vector of tags. Defaults to \code{"n"}
for each supplied word.}
}
\description{
Add one or more custom words to a jieba worker.
}
